+++ packages/sdk-core/src/constants.ts
@@ -1,6 +1,7 @@
 import DefaultSubgraphReleaseTag from "./defaultSubgraphReleaseTag.json";
 import { IResolverData } from "./interfaces";
 
+/******* TIME CONSTANTS *******/
 export const MONTHS_PER_YEAR = 12;
 export const DAYS_PER_MONTH = 30;
 export const DAYS_PER_WEEK = 7;
@@ -14,19 +15,89 @@ export const MONTH_IN_SECONDS = DAY_IN_SECONDS * DAYS_PER_MONTH;
 export const YEAR_IN_SECONDS = MONTH_IN_SECONDS * MONTHS_PER_YEAR; // NOTE: Is 360 days (misses 5-6 days)
 export const BASE_18 = 1e18;
 
-export const ROPSTEN = "ropsten";
-export const RINKEBY = "rinkeby";
-export const GOERLI = "goerli";
-export const KOVAN = "kovan";
-export const MUMBAI = "mumbai";
-export const OPTIMISM_KOVAN = "optimism-kovan";
-export const ARBITRUM_RINKEBY = "arbitrum-rinkeby";
-export const AVALANCHE_FUJI = "avalanche-fuji";
-export const GNOSIS = "gnosis";
-export const MATIC = "matic";
-export const OPTIMISM_MAINNET = "optimism-mainnet";
-export const ARBITRUM_ONE = "arbitrum-one";
+/******* NETWORK NAMES *******/
+export const ETH_ROPSTEN = "eth-ropsten";
+export const ETH_RINKEBY = "eth-rinkeby";
+export const ETH_GOERLI = "eth-goerli";
+export const ETH_KOVAN = "eth-kovan";
 
+export const POLYGON_MAINNET = "polygon-mainnet";
+export const POLYGON_MUMBAI = "polygon-mumbai";
+
+export const XDAI_MAINNET = "xdai-mainnet";
+
+export const OPTIMISM_MAINNET = "optimism-mainnet";
+export const OPTIMISM_KOVAN = "optimism-kovan";
+
+export const ARBITRUM_ONE = "arbitrum-one";
+export const ARBITRUM_RINKEBY = "arbitrum-rinkeby";
+
+export const AVALANCHE_C = "avalanche-c";
+export const AVALANCHE_FUJI = "avalanche-fuji";
+
+export const networkNames: string[] = [
+    ETH_ROPSTEN,
+    ETH_RINKEBY,
+    ETH_GOERLI,
+    ETH_KOVAN,
+
+    POLYGON_MAINNET,
+    POLYGON_MUMBAI,
+
+    XDAI_MAINNET,
+
+    ARBITRUM_ONE,
+    ARBITRUM_RINKEBY,
+
+    OPTIMISM_MAINNET,
+    OPTIMISM_KOVAN,
+
+    AVALANCHE_C,
+    AVALANCHE_FUJI,
+];
+
+/******* CHAIN IDS *******/
+export const ETH_ROPSTEN_CHAIN_ID = 3;
+export const ETH_RINKEBY_CHAIN_ID = 4;
+export const ETH_GOERLI_CHAIN_ID = 5;
+export const ETH_KOVAN_CHAIN_ID = 42;
+
+export const GNOSIS_CHAIN_ID = 100;
+
+export const MATIC_CHAIN_ID = 137;
+export const MUMBAI_CHAIN_ID = 80001;
+
+export const OPTIMISM_MAINNET_CHAIN_ID = 10;
+export const OPTIMISM_KOVAN_CHAIN_ID = 69;
+
+export const ARBITRUM_ONE_CHAIN_ID = 42161;
+export const ARBITRUM_RINKEBY_CHAIN_ID = 421611;
+
+export const AVALANCHE_FUJI_CHAIN_ID = 43113;
+export const AVALANCHE_C_CHAIN_ID = 43114;
+
+export const chainIds = [
+    ETH_ROPSTEN_CHAIN_ID, // ROPSTEN
+    ETH_RINKEBY_CHAIN_ID, // RINKEBY
+    ETH_GOERLI_CHAIN_ID, // GOERLI
+    ETH_KOVAN_CHAIN_ID, // KOVAN
+
+    MATIC_CHAIN_ID, // MATIC
+    MUMBAI_CHAIN_ID, // MUMBAI
+
+    GNOSIS_CHAIN_ID, // GNOSIS
+
+    OPTIMISM_MAINNET_CHAIN_ID, // OPTIMISM MAINNET
+    OPTIMISM_KOVAN_CHAIN_ID, // OPTIMISM KOVAN
+
+    ARBITRUM_ONE_CHAIN_ID, // ARBITRUM ONE
+    ARBITRUM_RINKEBY_CHAIN_ID, // ARBITRUM RINKEBY
+
+    AVALANCHE_FUJI_CHAIN_ID, // AVALANCHE FUJI
+    AVALANCHE_C_CHAIN_ID, // AVALANCHE C-CHAIN
+];
+
+/******* ACL AUTHORIZATION BIT OPERATIONS *******/
 export const AUTHORIZE_FLOW_OPERATOR_CREATE = 1 << 0;
 export const AUTHORIZE_FLOW_OPERATOR_UPDATE = 1 << 1;
 export const AUTHORIZE_FLOW_OPERATOR_DELETE = 1 << 2;
@@ -35,69 +106,39 @@ export const AUTHORIZE_FULL_CONTROL =
     AUTHORIZE_FLOW_OPERATOR_UPDATE |
     AUTHORIZE_FLOW_OPERATOR_DELETE;
 
-export const chainIds = [
-    3, // ROPSTEN
-    4, // RINKEBY
-    5, // GOERLI
-    10, // OPTIMISM MAINNET
-    42, // KOVAN
-    69, // OPTIMISM KOVAN
-    100, // GNOSIS
-    137, // MATIC
-    42161, // ARBITRUM ONE
-    43113, // AVALANCHE FUJI
-    80001, // MUMBAI
-    421611, // ARBITRUM RINKEBY
-];
-
-export const networkNames: string[] = [
-    ROPSTEN,
-    RINKEBY,
-    GOERLI,
-    KOVAN,
-    MUMBAI,
-    OPTIMISM_KOVAN,
-    ARBITRUM_RINKEBY,
-    AVALANCHE_FUJI,
-    GNOSIS,
-    MATIC,
-    OPTIMISM_MAINNET,
-    ARBITRUM_ONE,
-];
-
 const subgraphReleaseTag =
     process.env.SUBGRAPH_RELEASE_TAG || DefaultSubgraphReleaseTag.value;
 
 export const chainIdToResolverDataMap = new Map<number, IResolverData>([
     [
-        3,
+        ETH_ROPSTEN_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-ropsten`,
-            networkName: ROPSTEN,
+            networkName: ETH_ROPSTEN,
             resolverAddress: "0x3b44e06D96BcA9412CBc23F80F41B9e30933571a",
             nativeTokenSymbol: "ETH",
         },
     ],
     [
-        4,
+        ETH_RINKEBY_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-rinkeby`,
-            networkName: RINKEBY,
+            networkName: ETH_RINKEBY,
             resolverAddress: "0x659635Fab0A0cef1293f7eb3c7934542B6A6B31A",
             nativeTokenSymbol: "ETH",
         },
     ],
     [
-        5,
+        ETH_GOERLI_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-goerli`,
-            networkName: GOERLI,
+            networkName: ETH_GOERLI,
             resolverAddress: "0x3710AB3fDE2B61736B8BB0CE845D6c61F667a78E",
             nativeTokenSymbol: "ETH",
         },
     ],
     [
-        10,
+        OPTIMISM_MAINNET_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-optimism-mainnet`,
             networkName: OPTIMISM_MAINNET,
@@ -106,16 +147,16 @@ export const chainIdToResolverDataMap = new Map<number, IResolverData>([
         },
     ],
     [
-        42,
+        ETH_KOVAN_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-kovan`,
-            networkName: KOVAN,
+            networkName: ETH_KOVAN,
             resolverAddress: "0x851d3dd9dc97c1df1DA73467449B3893fc76D85B",
             nativeTokenSymbol: "ETH",
         },
     ],
     [
-        69,
+        OPTIMISM_KOVAN_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-optimism-kovan`,
             networkName: OPTIMISM_KOVAN,
@@ -124,25 +165,25 @@ export const chainIdToResolverDataMap = new Map<number, IResolverData>([
         },
     ],
     [
-        100,
+        GNOSIS_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-xdai`,
-            networkName: GNOSIS,
+            networkName: XDAI_MAINNET,
             resolverAddress: "0xD2009765189164b495c110D61e4D301729079911",
             nativeTokenSymbol: "xDAI",
         },
     ],
     [
-        137,
+        MATIC_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-matic`,
-            networkName: MATIC,
+            networkName: POLYGON_MAINNET,
             resolverAddress: "0xE0cc76334405EE8b39213E620587d815967af39C",
             nativeTokenSymbol: "MATIC",
         },
     ],
     [
-        42161,
+        ARBITRUM_ONE_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-arbitrum-one`,
             networkName: ARBITRUM_ONE,
@@ -151,7 +192,7 @@ export const chainIdToResolverDataMap = new Map<number, IResolverData>([
         },
     ],
     [
-        43113,
+        AVALANCHE_FUJI_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-avalanche-fuji`,
             networkName: AVALANCHE_FUJI,
@@ -160,16 +201,25 @@ export const chainIdToResolverDataMap = new Map<number, IResolverData>([
         },
     ],
     [
-        80001,
+        AVALANCHE_C_CHAIN_ID,
+        {
+            subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-avalanche-c`,
+            networkName: AVALANCHE_C,
+            resolverAddress: "0x24a3F04F70B7f07B9673EadD3e146391BcfEa5c1",
+            nativeTokenSymbol: "AVAX",
+        },
+    ],
+    [
+        MUMBAI_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-mumbai`,
-            networkName: MUMBAI,
+            networkName: POLYGON_MUMBAI,
             resolverAddress: "0x8C54C83FbDe3C59e59dd6E324531FB93d4F504d3",
             nativeTokenSymbol: "MATIC",
         },
     ],
     [
-        421611,
+        ARBITRUM_RINKEBY_CHAIN_ID,
         {
             subgraphAPIEndpoint: `https://api.thegraph.com/subgraphs/name/superfluid-finance/protocol-${subgraphReleaseTag}-arbitrum-rinkeby`,
             networkName: ARBITRUM_RINKEBY,
@@ -180,16 +230,22 @@ export const chainIdToResolverDataMap = new Map<number, IResolverData>([
 ]);
 
 export const networkNameToChainIdMap = new Map<string, number>([
-    [ROPSTEN, 3],
-    [RINKEBY, 4],
-    [GOERLI, 5],
-    [KOVAN, 42],
-    [OPTIMISM_KOVAN, 69],
-    [AVALANCHE_FUJI, 43113],
-    [MUMBAI, 80001],
-    [ARBITRUM_RINKEBY, 421611],
-    [GNOSIS, 100],
-    [MATIC, 137],
-    [OPTIMISM_MAINNET, 10],
-    [ARBITRUM_ONE, 42161],
+    [ETH_ROPSTEN, ETH_ROPSTEN_CHAIN_ID],
+    [ETH_RINKEBY, ETH_RINKEBY_CHAIN_ID],
+    [ETH_GOERLI, ETH_GOERLI_CHAIN_ID],
+    [ETH_KOVAN, ETH_KOVAN_CHAIN_ID],
+
+    [XDAI_MAINNET, GNOSIS_CHAIN_ID],
+
+    [POLYGON_MAINNET, MATIC_CHAIN_ID],
+    [POLYGON_MUMBAI, MUMBAI_CHAIN_ID],
+
+    [OPTIMISM_KOVAN, OPTIMISM_KOVAN_CHAIN_ID],
+    [OPTIMISM_MAINNET, OPTIMISM_MAINNET_CHAIN_ID],
+
+    [ARBITRUM_ONE, ARBITRUM_ONE_CHAIN_ID],
+    [ARBITRUM_RINKEBY, ARBITRUM_RINKEBY_CHAIN_ID],
+
+    [AVALANCHE_C, AVALANCHE_C_CHAIN_ID],
+    [AVALANCHE_FUJI, AVALANCHE_FUJI_CHAIN_ID],
 ]);