+++ packages/sdk-core/src/index.ts
@@ -8,10 +8,11 @@ import Governance from "./Governance";
 import Host from "./Host";
 import InstantDistributionAgreementV1 from "./InstantDistributionAgreementV1";
 import Query from "./Query";
-import SuperToken from "./SuperToken";
-import { WrapperSuperToken } from "./SuperToken";
-import { PureSuperToken } from "./SuperToken";
-import { NativeAssetSuperToken } from "./SuperToken";
+import SuperToken, {
+    NativeAssetSuperToken,
+    PureSuperToken,
+    WrapperSuperToken,
+} from "./SuperToken";
 
 export * from "./interfaces";
 export * from "./utils";
@@ -46,6 +47,7 @@ export * from "./subgraph/events/events";
 export * from "./subgraph/events/flowUpdatedEvent";
 export * from "./subgraph/events/indexUpdatedEvent";
 export * from "./subgraph/events/subscriptionUnitsUpdatedEvents";
+export * from "./subgraph/events/transferEvents";
 
 export * from "./subgraph/meta/meta";