+++ packages/sdk-core/src/subgraph/entities/accountTokenSnapshot/accountTokenSnapshot.ts
@@ -31,6 +31,8 @@ export interface AccountTokenSnapshot {
     totalNetFlowRate: BigNumber;
     totalNumberOfActiveStreams: number;
     totalOutflowRate: BigNumber;
+    maybeCriticalAtTimestamp: BigNumber;
+    isLiquidationEstimateOptimistic: boolean;
     totalNumberOfClosedStreams: number;
     totalSubscriptionsWithUnits: number;
     updatedAtBlockNumber: BlockNumber;