+++ packages/subgraph/src/mappings/idav1.ts
@@ -26,11 +26,13 @@ import {
 import {
     BIG_INT_ZERO,
     createEventID,
-    getIndexID, getOrder,
+    getIndexID,
+    getOrder,
     subscriptionExists as subscriptionWithUnitsExists,
     tokenHasValidHost,
 } from "../utils";
 import {
+    createAccountTokenSnapshotLogEntity,
     getOrInitIndex,
     getOrInitSubscription,
     getOrInitTokenStatistic,
@@ -77,6 +79,7 @@ export function handleIndexCreated(event: IndexCreated): void {
         event.block
     );
 
+    createAccountTokenSnapshotLogEntity(event, event.params.publisher, event.params.token, "IndexCreated");
     createIndexCreatedEntity(event, index.id);
 }
 
@@ -149,7 +152,7 @@ export function handleIndexUpdated(event: IndexUpdated): void {
         event.params.token,
         event.block
     );
-
+    createAccountTokenSnapshotLogEntity(event, event.params.publisher, event.params.token, "IndexUpdated");
     createIndexUpdatedEntity(event, index.id);
 }
 
@@ -242,6 +245,7 @@ export function handleSubscriptionApproved(event: SubscriptionApproved): void {
             event.params.token,
             event.block
         );
+        createAccountTokenSnapshotLogEntity(event, event.params.publisher, event.params.token, "SubscriptionApproved");
     }
 
     subscription.save();
@@ -263,6 +267,7 @@ export function handleSubscriptionApproved(event: SubscriptionApproved): void {
     index.save();
 
     createSubscriptionApprovedEntity(event, subscription.id);
+    createAccountTokenSnapshotLogEntity(event, event.params.subscriber, event.params.token, "SubscriptionApproved")
 }
 
 export function handleSubscriptionDistributionClaimed(
@@ -307,6 +312,8 @@ export function handleSubscriptionDistributionClaimed(
         event.params.token,
         event.block
     );
+    createAccountTokenSnapshotLogEntity(event, event.params.publisher, event.params.token, "SubscriptionDistributionClaimed");
+    createAccountTokenSnapshotLogEntity(event, event.params.subscriber, event.params.token, "SubscriptionDistributionClaimed");
 }
 
 /**
@@ -396,6 +403,8 @@ export function handleSubscriptionRevoked(event: SubscriptionRevoked): void {
     subscription.save();
 
     createSubscriptionRevokedEntity(event, subscription.id);
+    createAccountTokenSnapshotLogEntity(event, event.params.subscriber, event.params.token, "SubscriptionRevoked")
+    createAccountTokenSnapshotLogEntity(event, event.params.publisher, event.params.token, "SubscriptionRevoked");
 }
 
 /**
@@ -520,6 +529,8 @@ export function handleSubscriptionUnitsUpdated(
     subscription.save();
 
     createSubscriptionUnitsUpdatedEntity(event, subscription.id, oldUnits);
+    createAccountTokenSnapshotLogEntity(event, event.params.publisher, event.params.token, "SubscriptionUnitsUpdated");
+    createAccountTokenSnapshotLogEntity(event, event.params.subscriber, event.params.token, "SubscriptionUnitsUpdated");
 }
 
 /**************************************************************************